fix: psalm issues
This adds psalm and fixes all of the issues that come with that. The only change
in types is that `BaseDataset::$query` will now accept a `\yii\db\Command` this
is so you can use the `Yii::$app->db->createCommand()` method to create a query
to make into a chart.

Ref: #17